+++ src/components/common/sidebar/sidebar-content-footer.tsx
@@ -19,7 +19,7 @@ export function SidebarContentFooter() {
                         ? <SidebarContentListElement Icon={LogoutIcon} label={'Wyloguj'} link={''} onClick={onLogout} />
                         : <SidebarContentListElement Icon={LoginIcon} label={'Zaloguj'} link={'/users/login'} />
                 )}
-                <p className={styles['sidebar-content__header-app-version']}><i>Wersja - v.2.4.2</i></p>
+                <p className={styles['sidebar-content__header-app-version']}><i>Wersja - v.2.5.0</i></p>
             </div>
         </>
     );