+++ src/contexts/recipe-form.context.tsx
@@ -1,12 +1,18 @@
 'use client';
 
-import { createContext, useContext } from 'react';
+import { createContext, PropsWithChildren, useContext, useMemo } from 'react';
 import { DishRecipeSectionWithId } from '@/src/types/dish.types';
 import { FieldError, FieldErrorsImpl, Merge } from 'react-hook-form';
+import { invertTranslations, sanitizeTranslations } from '@/src/utils/recipe.utils';
+import { Translations } from '@/src/types/recipe.types';
 
-interface RecipeFormProviderProps {
-    children: any;
+import actionJson from '@/public/data/actions.json';
+import ingredientJson from '@/public/data/ingredients.json';
+import unitJson from '@/public/data/units.json';
+
+interface RecipeFormProviderProps extends PropsWithChildren {
     sections: DishRecipeSectionWithId[];
+    translations: Translations;
     onChangeSections: (sections: DishRecipeSectionWithId[]) => void;
     error?: Merge<FieldError, (Merge<FieldError, FieldErrorsImpl<DishRecipeSectionWithId>> | undefined)[]>;
 }
@@ -14,6 +20,11 @@ interface RecipeFormProviderProps {
 type RecipeFormContextValues = Omit<RecipeFormProviderProps, 'children'>;
 
 const defaultValue: RecipeFormContextValues = {
+    translations: {
+        actions: {},
+        ingredients: [],
+        units: {}
+    },
     sections: [],
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     onChangeSections: (sections: DishRecipeSectionWithId[]): void => {}
@@ -24,7 +35,15 @@ const RecipeFormContext = createContext<RecipeFormContextValues>(defaultValue);
 export const useRecipeFormContext = () => useContext(RecipeFormContext);
 
 export function RecipeFormProvider({ children, sections, onChangeSections, error }: RecipeFormProviderProps) {
-    const contextValue: RecipeFormContextValues = { sections, onChangeSections, error };
+    const translations = useMemo(() => {
+        const actions = invertTranslations(actionJson, 'pl');
+        const ingredients = sanitizeTranslations(ingredientJson);
+        const units = invertTranslations(unitJson, 'pl');
+
+        return { actions, ingredients, units };
+    }, []);
+
+    const contextValue: RecipeFormContextValues = { translations, sections, onChangeSections, error };
 
     return (
         <RecipeFormContext.Provider value={contextValue}>