+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.0] - 2025-04-23
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `actions`, `units` and moved `ingredients`.
+- [Marek Kurańda](https://github.com/mjkuranda): Utils and types for recipe part.
+
 ## [2.4.2] - 2025-04-16
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): Cut too long user login adding three dots at the end.